astrograph/requirements.txt
Flask
Click
numpy
opencv-python
pillow